02-DynamicModularity
Talks about adding services as a way of reacting to live services coming and going.
Note that this used both tinyurl and trim as shorteners; however, note that trim
has subsequently ceased operation.

Note also that Felix has changed command provider to a provisional Gogo shell;
this now includes an extra implemmentation in order to run in Felix Gogo.